# infrastructure/terraform/openapi.yaml
openapi: 3.0.3
info:
  title: Donation Platform API
  description: |
    Comprehensive API for the gamified donation platform with A2A agent integration.
    
    Features:
    - Donation processing with real-time gamification
    - Points and achievements system
    - A2A agent coordination for optimization
    - Location-based rewards and hotspots
    - Real-time analytics and leaderboards
    
  version: 1.0.0
  contact:
    name: Donation Platform Team
    email: support@donationplatform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: ${donation_service_url}
    description: Donation Service
  - url: ${points_service_url}
    description: Points & Achievements Service
  - url: ${donor_agent_url}
    description: Donor Engagement Agent (A2A)
  - url: ${charity_agent_url}
    description: Charity Optimization Agent (A2A)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  schemas:
...
    Achievement:
      type: object
      properties:
        id:
          type: string
          description: Unique achievement identifier
        name:
          type: string
          description: Achievement name
        description:
          type: string
          description: Achievement description
        type:
          type: string
          enum: [donation_milestone, location_discovery, streak_achievement, tier_progression, social_engagement]
        points:
          type: integer
          minimum: 0
          description: Points awarded for this achievement
        emoji:
          type: string
          description: Emoji representation
        unlocked_at:
          type: string
          format: date-time
          description: When the achievement was unlocked
        requirements:
          type: object
          description: Requirements to unlock this achievement
    LeaderboardEntry:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        name:
          type: string
          description: User display name
        points:
          type: integer
          description: Total points earned
        tier:
          type: string
          enum: [Bronze, Silver, Gold, Platinum, Diamond]
        rank:
          type: integer
          minimum: 1
          description: Current leaderboard position
        donations_count:
          type: integer
          description: Total number of donations made
        total_donated:
          type: number
          format: double
          description: Total amount donated in USD
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/points/award:
    post:
      summary: Award points to user
      description: Award points for various activities and check for new achievements
      tags:
        - Points & Achievements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - points
                - source
                - description
              properties:
                user_id:
                  type: string
                  format: uuid
                points:
                  type: integer
                  minimum: 1
                source:
                  type: string
                  enum: [donation, achievement, bonus, referral, social_share]
                description:
                  type: string
                  maxLength: 255
                metadata:
                  type: object
      responses:
        '200':
          description: Points awarded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  points_awarded:
                    type: integer
                  total_points:
                    type: integer
                  achievements_unlocked:
                    type: array
                    items:
                      $ref: '#/components/schemas/Achievement'
                  transaction_id:
                    type: string
  /api/v1/leaderboard/global:
    get:
      summary: Get global leaderboard
      description: Retrieve the global points leaderboard
      tags:
        - Leaderboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Leaderboard retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  leaderboard:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeaderboardEntry'
                  pagination:
                    type: object
  /api/v1/leaderboard/tier/{tier}:
    get:
      summary: Get tier-specific leaderboard
      description: Retrieve leaderboard for a specific user tier
      tags:
        - Leaderboard
      parameters:
        - name: tier
          in: path
          required: true
          schema:
            type: string
            enum: [Bronze, Silver, Gold, Platinum, Diamond]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 50
      responses:
        '200':
          description: Tier leaderboard retrieved successfully
        '400':
          description: Invalid tier specified
  /api/v1/users/{user_id}/achievements:
    get:
      summary: Get user achievements
      description: Retrieve all achievements unlocked by the user
      tags:
        - Points & Achievements
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User achievements retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user_id:
                    type: string
                    format: uuid
                  achievements:
                    type: array
                    items:
                      $ref: '#/components/schemas/Achievement'
                  total_count:
                    type: integer

  # Analytics API
...